# swagger.yaml
openapi: 3.0.0
info:
  description: “Micro-service for DO4M”
  version: “1.0.0"
  title: “ Hello MotherFuckers”
servers:
  - url: http://localhost:3000/
...
paths:
...
  /mission/{missionId}:
    get:
      tags:
        - mission
      summary: Find mission by ID
      description: Returns a single mission
      operationId: findById
      parameters:
        - name: missionId
          in: path
          description: ID of mission to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mission'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Mission'
        '400':
          description: Invalid ID supplied
        '404':
          description: Mission not found
    delete:
      tags:
        - mission
      summary: Remove a mission 
      description: Remove a mission 
      operationId: delete
      requestBody:
        description: Delete an existing mission
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mission'
          application/xml:
            schema:
              $ref: '#/components/schemas/Mission'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Mission'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
      security:
        - petstore_auth:
            - write:missions
            - read:missions
...
components:
  schemas:
...
    Mission:
      required:
        - id
        - picture
        - status
        - price
        - title 
        - description
        - creation_date
        - id_create 
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        picture:
          type: string
        status:
          type: string
          description: mission status
        title:
          type: string
          example: Garder mon chien  
        price:
          type: number
          example: 35
        description:
          type: string
          example: Je cherche une dog sitter pour s'occuper de mon chien le 24 décembre
        creation_date : 
          type : date
          example : "2022-04-09T23:00:00.000Z"
      xml:
        name: mission
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://localhost:3000.swagger.io/oauth/authorize
          scopes:
            write:pets: modify missions in your account
            read:pets: read your missions
    api_key:
      type: apiKey
      name: api_key
      in: header